Applied Eclipse's cleanup function on all projects with very conservative settings.
git-svn-id: https://svn.jfire.org/svn/main/trunk/org.nightlabs.jfire.base.ui@9055 18c61ebf-f106-0410-9e1e-81fb82415b8f

+++ src/org/nightlabs/jfire/base/ui/prop/edit/blockbased/ImageDataFieldEditor.java
@@ -159,7 +159,7 @@ extends AbstractDataFieldEditor<ImageDataField>
 						messageBoxImageExceedsMaxSizeKB.setMessage(
 								String.format(
 										Messages.getString("org.nightlabs.jfire.base.ui.prop.edit.blockbased.ImageDataFieldEditor.messageBoxImageExceedsMaxSizeKB.message"), //$NON-NLS-1$
-										new Object[] { new Long(imageStructField.getMaxSizeKB()), new Long((long)(file.length() / 1024)) })
+										new Object[] { new Long(imageStructField.getMaxSizeKB()), new Long((file.length() / 1024)) })
 						);
 						messageBoxImageExceedsMaxSizeKB.open();
 						return;
